+++ spec/support/entities/computer_hardware.rb
@@ -1,0 +1,8 @@
+class ComputerHardware
+  include Datamappify::Entity
+
+  attribute :cpu, String
+  attribute :ram, Integer
+  attribute :hdd, Integer
+  attribute :gfx, String
+end