+++ spec/support/entities/computer_software.rb
@@ -1,0 +1,9 @@
+class ComputerSoftware
+  include Datamappify::Entity
+
+  attribute :os, String
+
+  references :osx
+  references :windows
+  references :linux
+end